improved accuracy
I fixed one of the bugs that could cause a "Mark4 not found" error
to occur, I'm not sure if there are any other triggers to this.

Additionally I fixed the out.bin outfile, it assumed that structs
would always be padded the same on different systems, which of course
isn't true.